Add animation effect to stat widgets + split code to improve readability
